# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The task - Rest API (Laravel 5.7)",
        "description": "Формат ответа на запрос <br><br> ``` {'success': true, 'message': 'Успешный ответ', 'data': null, 'code' :200}  ```",
        "contact": {
            "email": "balymbetov.temirlan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rest.test/api",
            "description": "Task - REST API OpenApi Server"
        }
    ],
    "paths": {
        "/towns": {
            "get": {
                "tags": [
                    "Города"
                ],
                "summary": "Список всех городов",
                "description": "Возвращает список всех городов",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Cоздание пользователя",
                "description": "Создание пользователя",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "surname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Список всех пользователей",
                "description": "Возвращает список всех пользователей",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Информация о пользователя",
                "description": "Возвращает информацию о пользователя",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/user/count": {
            "get": {
                "tags": [
                    "Количество пользователей"
                ],
                "summary": "Количество пользователей",
                "description": "Возвращает количество пользователей",
                "operationId": "getCountUsers",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}